+++ src/hooks/auth/useAuthSession.ts
@@ -1,8 +1,9 @@
 import { useEffect, useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { clearClientSession } from '@apis/auth/sessionActions';
+import { restoreSession } from '@apis/auth/sessionRestore';
 import useAuthStore from '@stores/authStore';
 import useSocketStore from '@stores/socketStore';
-import { clearClientSession, restoreSession } from '@apis/authSession';
 import { PATH } from '@constants/path';
 import { AUTH_RESTORE_DISABLED_KEY } from '@constants/storage';
